download homologset directly from entrez
