+++ phylogenetics/exttools/README.md
@@ -1,0 +1,4 @@
+# External Dependencies
+
+The following dependencies must be downloaded to access the full
+power of the `phylogenetics` package.
